requirements.txt
boris_spider==1.0.6
